# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Numerical",
      "version": "1.0.0",
      "description": "API "
    },
    "servers": [
      {
        "url": "http://localhost:4000"
      }
    ],
    "paths": {
        "/api/equations": {
            "get": {
              "description": "",
              "tags": ["Get Equations"],
              "responses": {
                "200": {
                  "description": "OK"
                },
                "404": {
                  "description": "Not Found"
                }
              }
            }
          },
      "/api/Add-equations": {
        "post": {
          "summary": "Add a new equation for root-finding",
          "tags": ["Add Equations"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "methodType": {
                      "type": "string",
                      "description": "MethodType",
                      "example": "Bisection"
                    },
                    "func": {
                      "type": "string",
                      "description": "equation",
                      "example": "x^2 - 4"
                    },
                    "xL": {
                      "type": "number",
                      "description": "XL",
                      "example": 1
                    },
                    "xR": {
                      "type": "number",
                      "description": "XR",
                      "example": 2
                    },
                    "table": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "xl": { "type": "number" },
                          "xr": { "type": "number" }
                        }
                      },
                      "description": "Iteration table",
                      "example": [{"iteration":2 ,"xl": 1, "xr": 2 ,"error":20}]
                    },
                    "epsilon": {
                      "type": "number",
                      "description": "epsilon",
                      "example": 0.01
                    },
                    "answer": {
                      "type": "number",
                      "description": "Answer",
                      "example": 2
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Equation added successfully"
            },
            "500": {
              "description": "Failed to save equation"
            }
          }
        }
      },
      "/api/Integral": {
        "get": {
          "description": "",
          "tags": ["Get Equations Integral"],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            }
          }
        }
      },
      "/api/Add-Integral": {
        "post": {
          "summary": "Add a new equation for Integral",
          "tags": ["Add Integral"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "methodType": {
                      "type": "string",
                      "description": "MethodType",
                      "example": "Trapezoidal"
                    },
                    "equation": {
                        "type": "string",
                        "description": "Mathematical equation to integrate",
                        "example": "x^2-4"
                      },
                      "a": {
                        "type": "number",
                        "description": "Lower bound of integration",
                        "example": 1
                      },
                      "b": {
                        "type": "number",
                        "description": "Upper bound of integration",
                        "example": 2
                      },
                      "n": {
                        "type": "number",
                        "description": "Number of subdivisions",
                        "example": 2
                      },
                      "answer1": {
                        "type": "number",
                        "description": "First answer (e.g., calculated area)",
                        "example": 2
                      },
                      "answer2": {
                        "type": "number",
                        "description": "Second answer (if applicable)",
                        "example": 2
                      },
                      "err": {
                        "type": "number",
                        "description": "Estimated error",
                        "example": 20
                      }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Equation added successfully"
            },
            "500": {
              "description": "Failed to save equation"
            }
          }
        }
      },
      "/api/Diff": {
        "get": {
          "description": "",
          "tags": ["Get Differential"],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            }
          }
        }
      },
      "/api/Add-Diff": {
        "post": {
          "summary": "Add a new equation for Differential",
          "tags": ["Add Differential"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "methodType": {
                        "type": "string",
                        "description": "MethodType",
                        "example": "Fowardh"
                      },
                      "equation": {
                        "type": "string",
                        "description": "Equation",
                        "example": "e^(x+2)"
                      },
                      "x": {
                        "type": "number",
                        "description": "x value",
                        "example": 0
                      },
                      "h": {
                        "type": "number",
                        "description": "h value",
                        "example": 0.1
                      },
                      "degree": {
                        "type": "number",
                        "description": "Degree",
                        "example": 2
                      },
                      "answer1": {
                        "type": "number",
                        "description": "answer",
                        "example": 1.5
                      },
                      "answer2": {
                        "type": "number",
                        "description": "exact answer",
                        "example": 2.5
                      },
                      "err": {
                        "type": "number",
                        "description": "err",
                        "example": 0.01
                      }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Equation added successfully"
            },
            "500": {
              "description": "Failed to save equation"
            }
          }
        }
      },
      "/api/Gauss": {
        "get": {
          "description": "",
          "tags": ["Get Gauss"],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            }
          }
        }
      } ,
      "/api/Add-Gauss": {
        "post": {
          "summary": "Add a new equation for Gauss",
          "tags": ["Add Gauss"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "methodType": {
                        "type": "string",
                        "description": "MethodType",
                        "example": "Gauss"
                      },
                    "equation": {
                        "type": "array",
                        "description": "Array ",
                        "example": [[2, 1, -1, 8], [ -3, -1, 2, -11], [ -2, 1, 2, -3]]
                      },
                      "size": {
                        "type": "number",
                        "description": "Size ",
                        "example": 3
                      },
                      "answer": {
                        "type": "array",
                        "description": "Array of solutions",
                        "example": [2, 3, -1]
                      },
                      "err": {
                        "type": "number",
                        "description": "error",
                        "example": 0.01
                      },
                      "table": {
                        "type": "array",
                        "description": "table for seidal or jacob",
                        "example": [{"x1":50,"x2":100,"x3":200}]
                      }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Equation added successfully"
            },
            "500": {
              "description": "Failed to save equation"
            }
          }
        }
      },
      "/api/Interpolation": {
        "get": {
          "description": "",
          "tags": ["Get Intepolation"],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            }
          }
        }
      },
      "/api/Add-Interpolation": {
        "post": {
          "summary": "Add Interpolation",
          "tags": ["Add Interpolation"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "methodType": {
                        "type": "string",
                        "description": "Methodtype ",
                        "example": "Lagrange"
                      },
                      "points": {
                        "type": "array",
                        "description": "Data points for interpolation",
                        "example": [[1, 2], [2, 3], [3, 5]]
                      },
                      "xTarget": {
                        "type": "number",
                        "description": "Target x ",
                        "example": 2.5
                      },
                      "answer": {
                        "type": "number",
                        "description": "answer",
                        "example": 2.75
                      },
                      "n": {
                        "type": "number",
                        "description": "Number of data points",
                        "example": 3
                      },
                      "chart": {
                        "type": "array",
                        "description": "Chart data ",
                        "example": [[1, 2], [2, 3], [3, 5], [2.5, 2.75]]
                      }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Equation added successfully"
            },
            "500": {
              "description": "Failed to save equation"
            }
          }
        }
      },
      "/api/Regression": {
        "get": {
          "description": "",
          "tags": ["Get Regression"],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            }
          }
        }
      } ,
      "/api/Add-Regression": {
        "post": {
          "summary": "Add Regression",
          "tags": ["Add Regression"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "methodType": {
                        "type": "string",
                        "description": "Method type ",
                        "example": "MultipleRegression"
                      },
                      "result": {
                        "type": "number",
                        "description": "Result ",
                        "example": 0.95
                      },
                      "regressionEquation": {
                        "type": "string",
                        "description": "Regression equation",
                        "example": "y = 2 + 3*X1 + 4*X2"
                      },
                      "points": {
                        "type": "array",
                        "description": "Data points ",
                        "example": [[1, 2], [2, 3], [3, 5]]
                      },
                      "plotData": {
                        "type": "array",
                        "description": "Data points for ",
                        "example": [[1, 1.5], [2, 2.5], [3, 4.5]]
                      },
                      "X1target": {
                        "type": "number",
                        "description": "Target  X1",
                        "example": 2
                      },
                      "X2target": {
                        "type": "number",
                        "description": "Target  X2",
                        "example": 3
                      }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Equation added successfully"
            },
            "500": {
              "description": "Failed to save equation"
            }
          }
        }
      }      
      
    }
  }
  